# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HireForge AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: jobs
    description: Job tracking
  - name: resumes
    description: Resume analysis
  - name: github
    description: GitHub profile analysis
  - name: reports
    description: Analytics and reports
  - name: dashboard
    description: Dashboard aggregates
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List jobs for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Add a job to tracker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job status or details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Remove job from tracker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /jobs/stats:
    get:
      operationId: getJobStats
      tags: [jobs]
      summary: Job status breakdown counts
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStats"

  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List analyzed resumes
      responses:
        "200":
          description: Resumes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"

  /resumes/analyze:
    post:
      operationId: analyzeResume
      tags: [resumes]
      summary: Submit resume for AI analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get resume analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

  /github/analyze:
    post:
      operationId: analyzeGithub
      tags: [github]
      summary: Analyze a GitHub profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubInput"
      responses:
        "201":
          description: GitHub analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubAnalysis"

  /github/analyses:
    get:
      operationId: listGithubAnalyses
      tags: [github]
      summary: List past GitHub analyses
      responses:
        "200":
          description: Analyses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GithubAnalysis"

  /github/analyses/{id}:
    get:
      operationId: getGithubAnalysis
      tags: [github]
      summary: Get a GitHub analysis by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GitHub analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GithubAnalysis"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List generated reports
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Dashboard KPI summary
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/activity:
    get:
      operationId: getDashboardActivity
      tags: [dashboard]
      summary: Recent activity feed
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Platform-wide admin stats
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        createdAt:
          type: string

    UserUpdate:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        title:
          type: string
        avatarUrl:
          type: string

    Job:
      type: object
      required: [id, title, company, status, appliedAt, userId]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        location:
          type: ["string", "null"]
        status:
          type: string
        salary:
          type: ["string", "null"]
        url:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        appliedAt:
          type: string
        userId:
          type: integer

    JobInput:
      type: object
      required: [title, company, status]
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        status:
          type: string
        salary:
          type: string
        url:
          type: string
        notes:
          type: string
        appliedAt:
          type: string

    JobUpdate:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        status:
          type: string
        salary:
          type: string
        url:
          type: string
        notes:
          type: string

    JobStats:
      type: object
      required: [total, applied, screening, interview, offer, rejected]
      properties:
        total:
          type: integer
        applied:
          type: integer
        screening:
          type: integer
        interview:
          type: integer
        offer:
          type: integer
        rejected:
          type: integer

    Resume:
      type: object
      required: [id, fileName, score, userId, analyzedAt]
      properties:
        id:
          type: integer
        fileName:
          type: string
        score:
          type: integer
        summary:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        userId:
          type: integer
        analyzedAt:
          type: string

    ResumeInput:
      type: object
      required: [fileName, content]
      properties:
        fileName:
          type: string
        content:
          type: string

    GithubAnalysis:
      type: object
      required: [id, username, score, userId, analyzedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        score:
          type: integer
        repoCount:
          type: integer
        commitScore:
          type: integer
        topLanguages:
          type: array
          items:
            type: string
        summary:
          type: ["string", "null"]
        strengths:
          type: array
          items:
            type: string
        userId:
          type: integer
        analyzedAt:
          type: string

    GithubInput:
      type: object
      required: [username]
      properties:
        username:
          type: string

    Report:
      type: object
      required: [id, title, type, createdAt, userId]
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
        summary:
          type: ["string", "null"]
        metrics:
          type: object
          additionalProperties: true
        userId:
          type: integer
        createdAt:
          type: string

    DashboardStats:
      type: object
      required: [totalJobs, activeApplications, resumeScore, githubScore, weeklyActivity]
      properties:
        totalJobs:
          type: integer
        activeApplications:
          type: integer
        resumeScore:
          type: integer
        githubScore:
          type: integer
        weeklyActivity:
          type: integer
        offersReceived:
          type: integer
        interviewsScheduled:
          type: integer

    ActivityItem:
      type: object
      required: [id, type, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        createdAt:
          type: string

    AdminStats:
      type: object
      required: [totalUsers, totalJobs, totalResumes, totalGithubAnalyses, activeToday]
      properties:
        totalUsers:
          type: integer
        totalJobs:
          type: integer
        totalResumes:
          type: integer
        totalGithubAnalyses:
          type: integer
        activeToday:
          type: integer
        newUsersThisWeek:
          type: integer