# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  totalJobs: number;
  totalResumes: number;
  totalGithubAnalyses: number;
  activeToday: number;
  newUsersThisWeek?: number;
}