# lib/api-zod/src/generated/types/githubAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GithubAnalysis {
  id: number;
  username: string;
  score: number;
  repoCount?: number;
  commitScore?: number;
  topLanguages?: string[];
  /** @nullable */
  summary?: string | null;
  strengths?: string[];
  userId: number;
  analyzedAt: string;
}