# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HireForge AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobUpdate {
  title?: string;
  company?: string;
  location?: string;
  status?: string;
  salary?: string;
  url?: string;
  notes?: string;
}